Update PATCH /projects/:id/backlog to add a "remove" action
This is part of story #10143 add/remove backlog items in planning view

PATCH projects/:id/backlog now allows a new optionnal parameter named "remove":
{
  "remove" : [
    {"id": 143},
    ...
  ]
}

An error is thrown if this option is used in classic backlog management.

An error is thrown if at least one of the provided artifact id to remove
from the project top backlog is not in the explicit top backlog.

Change-Id: Iab7acb20e1e01ee7527fe2c75802cb89381f6b74