Display different download agreement on packages
In order to test you will need to feed your database manually like:
- insert into frs_download_agreement (project_id, title, content) VALUES
(101, "Custom license agreement", "Lorem Ipsum is simply dummy text)
- insert into frs_package_download_agreement(package_id, agreement_id) VALUES (470, 1);

Where 101 in the first SQL statement is the id of the project where you
want to run your tests.
And 470 is the id of a package on which you want to run tests on. Please note
that this package must have "Yes" set to "License Approval".

Given this, the one package with custom agreement should display the text from the database
and allow to approve (and download) or decline.
Other packages that are using the default license agreement should get
the standard modal.

Part of story #14008 customize FRS Download Agreement in my project

Change-Id: I2ddc6c563626eeff104641f1f12cd033d633c8aa