story #14088 Use 'StatusOpen' instead of 'StatusAll' to get all current releases
Before, in API when we want all current releases with 'project/$id/milestone' {"period":"current"},
there was a treatment on all releases (Open and Closed), but in AD there is only open releases are displayed.

Now, to be faster, there is a treatment only on open releases (like AD), and get only open current releases.

How to test:
	- Have a release with no status
	- Have a release with status closed
	- Have a release with status opened and no dates
	- Have a release with status opened and dates
	- Go to API -> project/$id/milestone with query {"period":"current"}

Expected result:
	- Get only open current releases.

Change-Id: I2d3733dbe6d6859d8541cfa23cd9cbae6759005b