Add add/remove in actions dropdown in artifact view
This is part of story #13998 add/remove backlog items in artifact view

If an artifact can be added into the explicit top backlog, then the add
action is displayed into artifact actions.

If an artifact is in the explicit top backlog, then the remove action is
displayed into the artifact actions.

Both actions are displayed, we cannot click on them for now.
Currently, there is no glyph for these two actions. They will be added
into a dedicated commit.

Change-Id: I90923ecba4bf1540b732c55c2dfff6cd11b1a0bb