story #14093 Not display the good color at the beginning of a release
In plugin release widget, when we start a release, we have initial effort == remaining effort.
But, in the plugin, when we have the same points, it meant that the release was finished, but it was not true, and the points are not displayed with the good color.

Now, there is the good color on the points.

How to test:
	- Have some relases with:
		- remaining effort == initial effort
		- remaining effort > initial effort
		- remaining effort == 0
		- remaining effort < initial effort

Expect results:
	- Remaining effort == initial effort
		-> color = $tlp-ui-success
	- remaining effort > intiial effort
		-> color = $tlp-ui-dimmed
	- remaining effort == 0
		-> $tlp-theme-color
	- remaining effort < initial effort
		-> $tlp-ui-success

Change-Id: I36de8e69ba672c5d6d77124e910fbe5cb0c27ef9