Show/hide closed items is saved in user prefs
By default, closed items are displayed.

If user choose to hide closed items for a milestone taskboard,
Then on refresh, the closed items are still hidden.

If user choose to display closed items for a milestone taskboard,
Then on refresh, the closed items are still displayed.

Part of story #13799: Expand / collapse columns

Change-Id: I05c3af8a39619ec0bc537ede7a26b7f3ec808950