story #14175 Create empty burndown
To continue to include burndown graph in plugin release widget, there is now an empty burndown graph.

There is only 3 lines (lines scale and ideal line).
The dimension of the graph (width and height) are the same ratio (height/width) of burndown sprint in agiledashboard (height = 45% * width).

How to test:
	- Have some releases with burndown

Expected result:
	- In plugin release widget, when you toggle a release, you see an empty burndown graph at the right side if it exists.
	- If the burndown is under calculation, there is a message
	- If there is an error (start date or duration), there is an error

Change-Id: Ie663811ac5bda20e416754dcf7fd754e01338dfe