request #14094 All groups members should display all its members
When an user is project member and project administrator with  the deleted status, it was not possible to remove him from the project
This patch goal is to display of the suspended and deleted users in a group member regardless the group type (dynamic or static).
It will allow project administrators to remove the "deleted" user from the project by removing him to the administrator group. So the site administrator is not needed anymore to remove a deleted a project member user who was also project administrator.

How to test:
 - Without this patch
 - Add an user who will be deleted in a project as a project member
 - Add this user in the administrator group
 -  As a site admin, put the status of the user to "Deleted"
=> you should not see the deleted users in your groups whatever the group type.
=> You cannot remove the remove the user from project

 - With the patch
=> you should see the deleted users in your groups
=> you can remove the deleted user from the Administrator group
=> Then you can remove the deleted user from the project member group

Change-Id: I8af8390dd653ac5fb5f21c16647b00b7058168b9