Display label + count in collapsed column
When a column is collapsed, the label and the total number of cards in
this column should be displayed when the column has hover.

Known issue: when there is horizontal scroll, and a collapsed column
start to disappear behind the swimlane header, then when we hover the
column, the label is displayed above everything. This is due to the fact
that the label is part of the taskboard head (sticky with high z-index
so that it is always visible when we vertical scroll). This will be
fixed later, when we will have enough resources to work on this
relatively small issue.

Part of story #13799: Expand / collapse columns

Change-Id: Ic8b471d2b277059d6ceb6d7475bde3d2491b6d88