Edit package with custom agreement
Manage the license agreements at package creation and update.

Rather a big patch as everything is done in one place and there are a lot
of configuration to take into account:
- License Agreement mandatory or no
- Default agreement
- Custom agreement or not

This patch also slightly change the old behaviour when agreement is mandatory.
Until now, a mandatory agreement meant that approval was always set and no options
were displayed to end users. However with custom license agreement, another agreement
can be selected so we have to deal with this configuration.

Part of story #14008 customize FRS Download Agreement in my project

Change-Id: I376c3e20b07bf88854f7245ca1b5a85ca3537d54