Templated value in project template used in create_test_env must be escaped
With the change introduced in 0b965b4eb8c9e16e299f8875734ee8e581cc8f52
the values replaced in the XML are not escaped at all which might lead
to security issues such as XXE. Due to some other validation the issue
does not seem to be exploitable but we have no guarantee it's going to
stay that way in the future.

While adding some tests, it has become visible that the validation done
on the values was also incorrect, the opportunity has been taken to fix
that as well.

Part of request #14048: Slowness at artifact creation

Change-Id: I3554d96a94ef807dfb660bd58f1e3455bc9f8b3b