Edit remaining effort on the taskboard
Given a parent card with a remaining effort that I have permission to
edit,
When I click on the remaning effort badge,
Then it switches to edit mode
And I can enter a new value.

Given I hit [enter]
Then it saves the new value
And if I reload the page the new value is displayed.

Given I hit [escape]
Then it cancel the edition.

This contribution is considered to be good enough to be able to edit the
remaining effort during a daily standup. On the UX side it may be
lacking some visual indication on how to save/cancel the remaining
effort, but it can be done later after some dogfooding.

Part of story #14115: update the remaining effort of a parent card

Change-Id: Ie76f59ad91dbae7ca57f6be0ccc8270d17e35cbb