Refactoring: event for additional button is thrown at tracker_artifact level
This is part of story #10143 add/remove backlog items in planning view

No functional change expected here.
This is done because the add/remove to explicit backlog will need to
include a javascript file. It must be done at tracker_artifact level
to be able to call $GLOBALS['HTML']->includeFooterJavascriptFile();

Change-Id: Iabe8f4630cae9ff2f5554cb10eda6786575ccb10