request #14106 Update to Cypress v3.6.0
How to test:
- make tests_cypress should run all tests and succeed
- make tests_cypress_dev should run a Tuleap instance with seeded data.
Please follow the developper guide to run tests in development.

I took this opportunity to remove the "test" image for the "dev" make
target. We always use "npx" to run cypress, which means we never used
this image.
I also took the opportunity to upgrade the base cypress image tag to
node v12.

Change-Id: I21a0ceae82e5229fee8aaf02c1755a8a907e185d