Get Mapping without cardwall config
Part of story #13626 Drag and drop cards

This is a technical change. No functional change expected. The column
mappings JSON should be identical.

Why this refactoring?

> "Which field values are mapped to this column for this tracker in this
milestone's cardwall ?"

Before this refactoring, the only way to answer that question was to go
through the Cardwall_OnTop_Config for the milestone's tracker, which
always instanciates ALL columns and ALL their mappings. It can only give
us all mappings of a tracker, we then must filter it to get the mappings
for the column we care about.

This refactoring lets us be much more fine-grained and get the values
mapped to a single column. It will be heavily used to change the value
of cards for the REST route to move a card between columns. We must
check whether the user can update the mapped field and get the value_ids
mapped to the given column.

Without it, I would have to go through a long chain to get the mappings:
column_id => Cardwall_Column => Milestone => Planning => Config =>
Mappings => Mapping

With it, I can shorten this chain to:
column_id => Cardwall_Column => Mapping

Change-Id: Ifda562e7308f63e47b99e34fe5ae6126ba469f5e