Refactoring: Group Milestone Tracker and Tracker together
Part of story #13626 Drag and drop cards

This is a technical change. No functional change expected. The column
mappings JSON should be identical.

MilestoneTracker + Tracker are actually a sort of primary key when it
comes to mappings, so it makes sense to group them together in an
object. One less parameter.

Change-Id: I7cca16f6a0349f671e269797010d397daab9fc7a