request #14001: Update PrismJS (syntax highlighter) to 1.17.1
prism-languages does not work anymore (and is not really needed) [0]
so it has been removed and replaced by the equivalent components of PrismJS.

With the upgrade, the opportunity has been took to update the list of
supported languages for the syntax highlighting.
Added support of: HCL and JSX.
Removed support of: Brainfuck and lolcode.

[0] https://github.com/survivejs/prism-languages/issues/2

Change-Id: Idc0d44a569519a805010e76e8cccf15cdeab2acd