request #14142 Project admins might not be able to download some FRS files
Given an empty user group in a project, if only this user group is defined in the
permission of a package, and then a new release is created in this package,
files added in this release are visible through the web UI to the project
administrators but they are not able to download it.

Unit tests are not added because of the direct calls of permission_exist
and permission_is_authorized.

Change-Id: I7f057a3870627eb86962c1f02d58c671b9112621