Inject count of explicit topBacklog
This is part of story #14183 Inject Topbacklog count when it's an explicit backlog in plugin release Widget

When you had a project with an explicit backlog, the number of items displayed in plugin release widget (Roadmap Section)was not the good account.

Now, there is a difference to get this count if the project uses an explicit backlog or not.

How to test:
	- Have a project without an explicit backlog
	- Have some items in the backlog (the account displayed is correct in plugin release widget (Roadmap section))
	- Change the project to use explicit backlog
	- The account displayed in plugin release widget is correct
	- Add/Remove some items from explicit backlog
	- The account displayed in plugin is still correct.

Change-Id: I9e9c3546e70ec808134ffa64adfb155fe298b1ae