Display list of reviewers
Reviewers section is displayed only if the user has rights
to merge the pull request or if there's a least one reviewer.

A toolip is shown when you hover the avatar.

For now, Add/Edit button doesn't do anything.

Part of story #14190: get email notifications on pull requests - first stage

Change-Id: I3db4cb906c1dd3f58449f75420978c36a2f2661e