REST creation of projects with XML templates
At the moment only 'scrum' template is available for test.

This commit also change the constraints applied on the REST route.
Starting now, any one can use the REST route to create projects.
Until now, you had to be "REST Project Manager" to do that.

Tests were updated to reflect this change.

This commit also bring changes on Trove cat browse limit pseudo const
that was converted to an actual const. It was easier to do it this
way rather than spending hours to identify the env pollution in
phpunit tests.

Part of story #8283 Enhance project creation user experience

Change-Id: I6b5a294d189a471e22a53d7bee8186b48ba5e888