Remove tuleap-vue-components alias
This alias has been migrated to work with Jest to help
the migration from Karma/Jasmine. Outside of the test
environnement there is only two usages so it can safely be removed.

Using aliases requires to maintain in sync multiple configuration
(Typescript, Webpack and Jest) and debugging it is painful.

Part of request #13806: Use Jest as the default test framework for JS
unit tests

Change-Id: Ice3330c6172cb82b39e6f14a55bc4b211f457b97