Reviewers can be set on a pull-request via the REST API
A new REST endpoint is available for this operation:
PUT /pull_requests/{id}/reviewers.

Only someone with the ability to merge a PR can update the
list of reviewers.
Only users that can see the PR can be added to the list of
reviewers.

Part of story #14190: get email notifications on pull requests - first stage

Change-Id: Iafcbb4d8a265fc1ed8c25cdbef9b0a1cb1cb3483