Remove tlp_mocks_alias
Part of request #13806: Use Jest as the default test framework for JS
unit tests

This alias used to be needed for jasmine mocks of TLP. Now, we always
mock it through Jest, it should be removed.

Using aliases requires to maintain in sync multiple configuration
(Typescript, Webpack and Jest) and debugging it is painful.

Change-Id: Iab166c1096ea30f5303671c6f70d207c731e0510