Switch to edit mode
When I click on a card, it switches to edition mode:
* save/cancel buttons appear.
* Pencil icon disappear.
* It does not change the status of the remaining effort.

If the remaining effort was already in edition mode, then only one pair
of save/cancel buttons is displayed.

When I hit cancel button, then the card is not anymore in edition mode.
When I hit save button, then the card is not anymore in edition mode,
real action will be done later.

When the card is in edition mode, it is still draggable, but we will
remove that in a dedicated changeset.

When the card is dragged, the pencil icon is not displayed.

Part of story #13629: Update the title semantic of a card

Change-Id: I2f87e3d6efbcffe2de7af609d0b220712defbe3b