GET /pull_requests/{id}/reviewers only returns users able to access the PR
The REST endpoint to udpate the reviewers on a PR rejects users
that cannot access it. Having those ones in the retrieved list
will cause discrepancies when updating it. Also this might be confusing
for the users since they might expect a review from one reviewer that can
not access the code.

The situation where a user is in the list of reviewers but cannot access it
can only happen when permissions are changed.

Part of story #14190: get email notifications on pull requests - first stage

Change-Id: Ibce76dde52880667a71faf49f094f483130f9c79