The widget's label includes the name of the tracker backlog
This is part of story #14212 Use the tracker's label of backlog in the label of release Widget

To be more understandable, the label of release widget changes. And use the label of the backlog's tracker.
So, the labels of each widget are differents.

The label of the plugin in administration is now: "Milestone Widget".

The label of the plugin in dashboard is now: "{{ tracker's label }} Widget".

How to test:
	- Have a dashboard
	- Install the Widget (it's under category "Agile Dashboard"
	- Change the name of the tracker

Expected Results:
	- The label of the widget changes

Next step: Delete 'Release' before each milestones in the widget

Change-Id: I35da3e689b64967b318d48976c7fde3ff382df63