Esc key cancels the edition of the card
This is the same behavior than remaining effort edition: if card is in
edition and I hit [Esc], then the card goes back to read mode.

No there is only one listener on the document that dispatch the esc key
press event to a) dragndrop b) cancel buttons.

Part of story #13629: Update the title semantic of a card

Change-Id: I983f402452bec118b38488fbefb667f78b8b5092