Extract PR mergeability verification to limit code duplication
This code to do this verification already exist in two different
code sections and it will also be needed to determine if a user
has the rights to access to the reviewer autocompleter endpoint.
This contribution only moves the code in one place to avoid duplicating
it a third time. It is still not a great architecture because the check
can easily be forgotten by a developper but at the very least the verification
has more chances to be consistent between different actions.

No functionnal changes are expected

Part of story #14190: get email notifications on pull requests - first stage

Change-Id: I6a83a86d9df578366d392da8810732b9cc489978