(Dis)allow restricted users to create projects
Until now (old project creation) it was possible to configure that
with tweaking of (in)famous
/etc/tuleap/site-content/en_US/include/restricted_user_permissions.txt

There is now a dedicated configuration var for that.

This commit is a bit fat because the new conf is integrated properly
in ProjectRegistrationUserPermissionChecker that had to be refactored.

Part of story #8283: Enhance project creation user experience

Change-Id: I2022e2b9a5acf3ff4728017b8fa1c165d72d998d