Better UX when clearing reviewers
When you want to remove all reviewers, it was not
clear that you have to submit the form with a empty field.

Now, when you open the modal where there is at least one reviewer
and you empty clear the field, the label button change and a
message appears telling you the pull request won't have any reviewer.

Part of story #14190: get email notifications on pull requests - first stage

Change-Id: I379b9a4bfb053086561af1fa18e2b04007e26047