TemplateFactoryTest::testItReturnsScrumTemplateXML randomly fails
Depending on the order of the execution (some tests does not clean
after themselves) this test can fail.
This test attempts to load a XML file via simplexml_load_file() which
does not work when the loading of external entities is disabled [0].

The test that was not cleaning after itself has been cleaned and the
usages of simplexml_load_file() removed in the PHPUnit tests.

The PHPUnit extension used to debug the issue has been left in place
to make sure the issue is not introduced again.

Part of story #8283: Enhance project creation user experience

[0] https://bugs.php.net/bug.php?id=62577

Change-Id: I80222cc619aa6c84f30b558c9bdecf0a81891d6a