request #14223 Statistics disk usage time collection must be the same per project
When the system_event STATISTICS_DAILY is launched, it will collect disk usage per projects.
However, if one service takes several seconds to compute its value, the disk usage collected
inside the same project will have the same date and not the same time.
This will lead to an empty value in the disk usage table in the project
administration (graphs are well displayed).

Change-Id: Id3ac3172cacf47f517913503fb145319eb5d1ecf