Display textarea to edit label
When card is in edit mode, then a textarea is displayed instead of the
label so that we can edit it.

The textarea expands in height to fit its content. This is done with a
"mirror" (a hidden pre element) to know the height of the text.

Please note that this is not possible to save the new label for now.

Part of story #13629: Update the title semantic of a card

Change-Id: Ib4c01bbb3c72ce73924e7f08ec94a1aa8be0bdd7