Refresh parent card when changing status
part of story #13626 Drag and drop cards

How to test:
- When you change the status of a child (task), then its parent swimlane
card should be refreshed to take into account cross-tracker triggers.
- When you change the status of a solo card, then it should be refreshed
to take into account workflow post actions.

There are only a couple ways to visualize the refresh on taskboard at
the moment:
- define Initial Effort as a List field and create a trigger that
changes it
- define Remaining Effort as a List field and update it with a trigger
- set up a background color that changes based on a List field (for
example Status) and update it with a trigger

Change-Id: I00822aa2a589639292d7879dc4f78789b8dc5af4