Group together closing and closing post-action of a pull-request
This change group together the fact of closing a PR and the fact of
adding this information into the timeline the PR. Until now, it was the
developer responsability to remember that it needs to do such action after
closing a PR. Since we are going to add more things to do after the closing
of PR we need to make sure those post actions will not be forgotten and
code is not duplicated.

This is a technical change, no functional change is expected.

Part of story #14190: get email notifications on pull requests - first stage

Change-Id: I9edc22b43389c1e8c4ed7db45b7997b51f0b6c31