UI to save the new label
This commit prepare the field to save the new label of a card.

This is not a usual approach since here we are doing the frontend first,
and the backend will come later. This seems odd, but it is needed to
have rapid feedback on the UI. Therefore the REST API call is mocked
with a sleep(2s) so that we can test the various anymations.

1. Switch a card to edition mode. Hit [enter] immediately. Since the label
didn't change, the card comes back to read mode without needing to save
anything.

2. Switch a card to edition mode. Enter a new label. Hit [enter].
2.1. The card switch to read mode with the new label.
2.2. The card blinks in blue (ui-info) for 2s so that the user knows that save is ongoing.
2.3. The card stops blinking and pulse in green (ui-success) so that the
user knows that save is successful.
2.4. If there was an errror during the save, the error modal asking to
reload the page would be displayed)

3. Switch a card to edition mode. Enter a new label. Click Save button.
Same behavior than 2.

4. Switch a card to edition mode. Enter a new label. Hit [esc] or
click Cancel button. The card comes back to read mode and the former
label is displayed.

Part of story #13629: Update the title semantic of a card

Change-Id: If07bfc40f87052a4479e7ee4ff0bb4f14c9f8b2c