Keep the history of changes made on the reviewers of a pull-request
This change is technical, nothing functional changes are
visible for now.

The history of the changes made on the reviewers of a pull-request
is now kept. This will be useful to send the notifications (easy
replay of the event in case of failure) and to add this information
in the timeline of the pull request.

You can check that the changes are kept by taking a look at the content
of plugin_pullrequest_reviewer_change and plugin_pullrequest_reviewer_change_user
tables.

Part of story #14190: get email notifications on pull requests - first stage

Change-Id: I1fe91ef0b1b03757f3db90c7d7875d01122a4a96