Refresh dropped card and parent on change of status
part of story #13626 Drag and drop cards

How to test:
- When you change the status of a child (task), then it should be
refreshed. Its parent swimlane card should also be refreshed.
- When you change the status of a solo card, then it should be
refreshed.

For child cards, the only way to see the refresh that I'm aware of today
is to set up the background color based on status (or the mapped field).
When you change columns, the card should change background.
For parent card, you can do the same.

Change-Id: I6f054fd8be85f3753deb4a829eae01d8bec96bd8