Remove "show" card animation
Part of story #13626 Drag and drop cards

The setTimeout() and conditional to show this animation slowed
tremendously the initial render of taskboard. To get an idea, on a large
taskboard with >1000 cards, it crashed my browser tab (when previously
it would be render completely in under 10 seconds).

When we always set the class (to remove the slow conditionals), it
messes up drag and drop UX as the animation is played every time the
user moves their dragged card.

We don't have a plan to make it work yet, so in order to move forward,
it's better to have a faster taskboard than to keep this animation.

Change-Id: I35d01868bf6046f53e3aca19a42de22a0f1cccb3