Use CSS custom properties instead of Sass for-loops
Part of story #14249 Taskboard technical improvements

No functional change expected. It decreases the size of the generated
.css file from 92KiB to 50KiB.

Loops using mix() or darken() or the ease-out-background could not be
removed because they rely on SASS functions to alter the variables.
The "taskboard-card-*" and "taskboard-card-background-*" have also been
kept but we could find a clever way to use "tlp-swatch-*" class instead.

Change-Id: I7f0042ce742813556c922b4c37c46a4eac4eacc1