A line in the log at the debug level when a notification should be sent
This contribution puts in place the cornerstone of the architecture
to process notification coming from pull request events. For now,
the notifications are processed synchronously.

In order to deal with the fact that Tuleap will have to deal with
PR notifications for vastly different type of events (i.e. there is not
much data in common when sending a notification because a new reviewer
as been added vs when a PR is merged) an event bus [0] has been implemented
on top of PSR-14 [1]. For each type of events it is possible to define
what the content of the notification should be.

The usage of PSR-14 besides being a good fit for this could allow us to
map it to the global async event management.

Part of story #14190: get email notifications on pull requests - first stage

[0] https://en.wikipedia.org/wiki/Event-driven_architecture
[1] https://www.php-fig.org/psr/psr-14/

Change-Id: Ie240e0283ed00e2df28bbf6d7d76acd6ce98f5b2