REST project creation enforce categories checks
The REST route should apply the same constraints than regular
project creation regarding categories.

Actually, as the code is shared with the project admin update
of categories, the checks are more complete with the REST route
than what is possible with regular project creation. For instance
it's possible to create a project via REST with categories that
are not mandatory or it's possible to use category values of more
than 1 level of depth.

Part of story #8283 Enhance project creation user experience

Change-Id: Ia311acf24dd84d05daed1a61f5dc7ff30ebb20f3