A basic text notification is sent when a new reviewer is added to a PR
There is nothing fancy in the mail being sent for now, it is a plaintext
email.

Sending emails in reaction to an event is implemented through a strategy
pattern [0] to make possible to deal with different ways for different
events or to ease the introduction of new behaviors (for example logging
received events for debugging purposes or sending webhooks).

Part of story #14190: get email notifications on pull requests - first stage

[0] https://en.wikipedia.org/wiki/Strategy_pattern

Change-Id: I95a1da5a2208f6e0ffc6a0c7d4726809960cc6f9