Crash when an OpenID Connect provider has a NULL value as icon or color
It seems at some point it was possible to have NULL for icon or
color in the plugin_openidconnectclient_provider table. Now,
when you create or update a provider without an icon or color an
empty string is set in the DB.
Commit cbdecd29f123254da93c97283c15eddb1d710ecc introduced an interface
to represent common value hold by a provider configuration and enabled
strict typing on the concrete implementations. The interface does not expect
null for a color and string and as a result a fatal error is encountered when
this case happen.

Part of request #12832: Support Azure AD as an OpenID Connect provider

Change-Id: Ibc9f9b8771855bc7ff5a9b39829d8b5933494d06