Add new architecture to be able to parse plugin URLS
This is part of request #14079 User logs do not set correctly project_id

No functional changes expected here.
project_id from project dashboard URL must be still correctly retrieved.

This introduces a userlog access object that is built from the request.
It's directly logged in database if all the project information can be
retrieved in the request.

If the project is missing in the request, the object is stored in a cache object
so that a new information about the project send by the projectProviderEvent
can be used to be added to userlog access object.
(bye bye my immutable principle friend :'( )

If no project information can be retrieved, the access log is stored
without project information at the end of the PHP process.

Change-Id: I4091174389d936462f1badf54a584e65c1785541