Contextualize modal for adding a Azure AD based provider
The existing modal was still using the same terms than the
creation for providers following OIDC standards when it should
be clear the creation is about Azure AD.
Also a lot of HTML elements were re-using the same IDs which is
allowed, this contribution make sure IDs are no re-used between contexts.

For now, the creation modal for a new Azure AD provider is only visible
when the feature flag open_id_azure_option is set to 1.

Part of request #12832: Support Azure AD as an OpenID Connect provider

Change-Id: Ib3d6354006ae6e382a65a96bf6208af3a53d23eb