Add some Git and Tracker URLs for userlog project retriving
This is the last part of request #14079 User logs do not set correctly project_id

a lot of tracker and git URL are now set with the right project_id in
the userlog access.

No new URL will be added with this (crappy) architechture.
New ones will require a complete refactoring.

Change-Id: I394350adbe7d8cea3ba2ed57a6db229db60c0112