A notification is sent when a pull request is abandoned
This contribution comes with the introduction of the "owners"
of a pull request. Owners are the users creating the pull-request or
updating it and the users listed as reviewers.

They are the ones receiving a notification when the a pullrequest is
abandoned.

Part of story #14190: get email notifications on pull requests - first stage

Change-Id: I626266320f419a095f6591f6e933bd97e77cc472