Display card as soon as possible
No need to refresh the page when we add a card in place.

As soon as the corresponding artifact is created, we can insert in the
taskboard the new card, in is_being_saved state. The card has same
animations than when a card is edited.

Part of: story #14148 Add in place a child card

Change-Id: I0df83eeb0fa5d32da47ab1b6034ce4ac7b0d2f19