Allow add-in-place of children for swimlanes with invalid mapping
A solo card with an invalid mapping should let the user add children
cards in place (when possible).

For example, create a User story with Status field set to "Review".
Create a custom mapping and unselect "Review" value for User story
tracker in the "Review" column.

Part of: story #14148 Add in place a child card

Change-Id: Ia0c691ff2e274fc1ab80392ae141675a96277e81