Make PullRequestUpdaterTest always happy
Ultimately there is no good reason for those tests
to needs the DB, they probably should be rewritten as
pure unit tests but this requires to rewrite completly
the tests. Let's fix waht we have and clean up later if needed,
they have "worked" as is for 3 years so they seems ot be good enough.

Without this contribution the test suite fails when executed with the
seed 1575978189.

Change-Id: I5bfbbed564594976ed9e1567c98efd8db7d5befd