Do not add in place if status is not submitable
Given a child tracker with a status that cannot be submitted,
When you hover a column,
You should not be able to add in place a card for this tracker.

Furthermore, the add in place should be allowed if status field is
required, since we are submiting it.

In addition, the A.D template is updated so that tasks can be added in
place.

Part of: story #14148 Add in place a child card

Change-Id: I751b93ba53bebbaf6f9da1bc973fd85e1bab7ef5