Add curve in burndown
This is part of story #14177 finish burndown

Add curve in burndown chart in release widget.

How to test:
	- Have some releases

Expected results:
	- Have a green curve in the burndown. Same curve in AD
	- If you toggle several times a same release, there will be the same chart

In this patch, the function getFormattedDates will change too.
To test its importance, you can revert this function and see Release Widget's behaviour
(the curve will be shifted after each toggle).
No others expected changes with this change (in the AD, the burndown will be the same as before).

Change-Id: Ia22161b403ede81c5c5f7d91a6ac302640df10e7