request #14285: Document UI crashes under some undetermined conditions
Crash has been detected in one production instance and is due to
a TypeError. The production code with error looking like
e.folded_by_map[o] is undefined.

This seems to be caused by an incorrect change in
7937a1d2d4f190175835fa0274f5164fe5a5430a.

The section of the code is not tested and removing the code
section does not seem to make much of difference. Since it is
hard to write a test for something we do not understand, the
fix is to remove the impacted code section.

Change-Id: I68b68b2f6392abc5329c32663d414ac66973aff6