bugs #14288 No message under IE11 in Release Widget
The plugin Release Widget is not supported under IE11.
So, there aren't any rules in webpack.config.js to build the plugin for IE11.
But, the message 'The plugin is not supported under IE11' was displayed with Vue.Js,
but this message never appeared because the plugin was never build for IE11.

Now the message is displayed with PHP.

How to test:
	- Run Windows and go to IE11

Expect results:
	- There is now the message 'The plugin is not supported under IE11. Please use a more recent browser'

Change-Id: I92325785d4e4f3895fa97705247cf1ac554072bb