Show a spinner when the card begins to be created
Part of story #14286 Finish add in place

How to test:
- Simulate high network latency using dev tools
- When you're adding in place a card, after hitting "save", the "Save
button icon should now be a spinner. Both "save and cancel" buttons
should be disabled. "Enter" or "Esc" keys should be ignored.
- After the card is created, the buttons and keys are handled like
before.

Note: The Cancel and Save buttons behaviour is not changed for Label and
Remaining effort edit in place. Changes on those will be a separate
commit.

Change-Id: Id68e165eb45e4ad2c0f49e4d26052926f2a67640