Fix "in Top Backlog" to work correctly with explicit mode
This is part of story #14000 manage "en masse" backlog operations

When "Top backlog" is selected as a report criterion, then it must
correctly handle the explicit backlog mode.

This patch also fixes a bug than can occurs on the SQL request part of a
report. In some non trivial cases (linked to the combinaison of
GROUP_CONCAT and DISTINCT), the retrivied artifact ids are OK but the
corresping last_changeset_ids are wrong. This patch also fixes that.

Change-Id: I51a501d1abb7e9b929e6f0dcbce1471ed401f32c