Improve add in place UX for solo cards
On a solo card, the button should say explicitely that it will create a
child, else user may think that it will create a sibling.

The top margin is fixed, so that there is not too much white space
above the button in an empty column.

Part of story #14286: Finish add in place

Change-Id: I2a5e41a8c4c9023332c53c697163c97240e68622