Prevent Solo card from disappearing after drop
Part of story #14249 Taskboard technical improvements

- Move a solo card from "Todo" to "On going" column
- Move back the solo card back to the "Todo" column

Without this patch, the solo card will be removed from the DOM.
With this patch, it should stay in the DOM.

It seems there is some subtle quirk when the "Add in place" button was
added to the default slot of DropContainerCell. Somehow, it works fine
for ChildCards which are in a v-for, but CardWithRemainingEffort is
somehow deleted or moved around by dragula. Removing "Add in place" from
the <slot> seems to have changed something in Vue's rendering of these
components and solves the "disappearing" problem.

I took this opportunity to do some sanity-improving refactoring:
Added SoloSwimlaneCell
renamed CardWithChildren => ChildrenSwimlane
renamed ColumnWithChildren => ChildrenCell

Change-Id: Id8e60501ab741dd99db2686765c6797dad2a4268