Add "Unplanned" option to the "In Milestone"
This is part of story #14000 manage "en masse" backlog operations

If the project of the tracker uses the explicit backlog managment, then
a new value is selectable in the tracker report "In Milestone"
criterion: Unplanned.

This allows the user to retrieve artifacts that are not in the
explicit top backlog and not already planned in submilstones.

Change-Id: I5e1e92d381540a4d2901ea03fd56b5b002667b9f