request #14312: Remove message about corrupted document manager approval tables
A data corruption bug was fixed on the document manager approval table (see request #13118)
a while ago.
Unfortunately it was not possible to fix all the existing data
and a warning message was added on the remaining potentially corrupted tables.
Displaying this warning requires to pull more data than it should be needed
to determine if the table is affected or not and since then we expect that approval
tables that matter have been fixed.

Note that if for any reason there still a need of troubleshooting the annotation
on the tables is kept in the DB.

Change-Id: Ia3cd5c935654e01325eeda894f763e2a6b413867