story #14402 To be compliant with codepen
There are 2 changements with this patch to be compliant with codepen and figma in widget Project Milestones (see acceptace criteria to see there).

First, there is a bold circle around the artifact's number (when you open a release, at the left there is the number of each artifacts in a circle).

Second, before, the progress bar of dates was always red when the number of days until end was > 0. And tuleap theme color, when the release was finished.

Now, this progress bar is success color if the progress is less than 80%, otherwise the color is danger.
Of course, if there is not start date or end date, the color is disabled, and if it's the end of the release, the color is tuleap theme.

Change-Id: I7cf103664dbcc5be505975119a044bd794775dfc